pyproject.toml
[project]
name = "video-cut"
version = "0.1.0"
description = "Desktop GUI tool for precise video clip extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-mpv>=1.0.8",
    "pyside6>=6.10.2",
]
